agentcore/genai-loft-demo/pyproject.toml
[project]
name = "genai-loft-demo-session31"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-opentelemetry-distro>=0.12.2",
    "bedrock-agentcore==0.1.7",
    "bedrock-agentcore-starter-toolkit==0.1.22",
    "boto3==1.40.47",
    "botocore==1.40.47",
    "ddgs>=9.8.0",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "strands-agents>=1.15.0",
    "strands-agents-tools>=0.2.14",
    "streamlit>=1.51.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "demos", "infrastructure"]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
clean = "scripts:clean"
launch = "scripts:launch"
deploy = "scripts:deploy"
logs = "scripts:logs"
debug_streaming = "scripts.debug_streaming:main"
